add class structures for whoisxml dns, webrisk section models
